+++ src/i18n_check/check/key_formatting.py
@@ -135,11 +135,11 @@ def invalid_key_formats_check(
     )
     format_key_or_keys = "keys" if len(invalid_keys_by_format) > 1 else "key"
 
-    invalid_keys_by_format_error = f"""❌ key_formatting error: There {format_to_be} {len(invalid_keys_by_format)} i18n {format_key_to_be} not formatted correctly. Please reformat the following {format_key_or_keys}:\n\n{invalid_keys_by_format_string}"""
+    invalid_keys_by_format_error = f"""❌ key-formatting error: There {format_to_be} {len(invalid_keys_by_format)} i18n {format_key_to_be} not formatted correctly. Please reformat the following {format_key_or_keys}:\n\n{invalid_keys_by_format_string}"""
 
     if not invalid_keys_by_format:
         rprint(
-            "[green]✅ key_formatting success: All i18n keys are formatted correctly in the [/green]"
+            "[green]✅ key-formatting success: All i18n keys are formatted correctly in the [/green]"
         )
         rprint("[green]i18n-src file.[/green]")